requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
numpy
mediapipe>=0.10.0
opencv-python-headless
